Fix signed vs unsigned comparison test

+++ src/render/backend/qgraphicsutils_p.h
@@ -333,7 +333,7 @@ static const T *valueArrayFromVariant(const QVariant &v, int count, int tupleSiz
     QVariantList vList = v.toList();
     // Handles list of QVariant: usually arrays of float
     if (!vList.isEmpty()) {
-        for (int i = 0; i < vList.length() && (uint)i * offset < uniformValuesArray.size(); i++) {
+        for (int i = 0; i < vList.length() && uint(i) * offset < uint(uniformValuesArray.size()); i++) {
             const char *subBuffer = QGraphicsUtils::bytesFromVariant<T>(vList.at(i));
             memcpy(data + i * offset, subBuffer, offset);
         }
